refactor: consolidate analytics helpers
