+++ analytics/core/utils/__init__.py
@@ -1,4 +1,17 @@
 from .unicode_processor import UnicodeHelper
+from .results_display import (
+    _extract_counts,
+    _extract_security_metrics,
+    _extract_enhanced_security_details,
+    create_analysis_results_display,
+    create_analysis_results_display_safe,
+)
 
-__all__ = ["UnicodeHelper"]
-
+__all__ = [
+    "UnicodeHelper",
+    "_extract_counts",
+    "_extract_security_metrics",
+    "_extract_enhanced_security_details",
+    "create_analysis_results_display",
+    "create_analysis_results_display_safe",
+]
