+++ services/__init__.py
@@ -16,8 +16,8 @@ else:
     from .analytics.data_loader import DataLoader
     from .analytics.publisher import Publisher
     from .analytics.upload_analytics import UploadAnalyticsProcessor
-    from .analytics_generator import AnalyticsGenerator
-    from .analytics_processor import AnalyticsProcessor
+    from .analytics.generator import AnalyticsGenerator
+    from .analytics.processor import AnalyticsProcessor
     from .async_file_processor import AsyncFileProcessor
     from .chunked_analysis import analyze_with_chunking
     from .controllers.upload_controller import UploadProcessingController
