+++ services/analytics/generator.py
