+++ services/analytics_summary.py
@@ -4,7 +4,7 @@ from typing import Any, Dict
 
 import pandas as pd
 
-from .analytics_generator import AnalyticsGenerator
+from .analytics.generator import AnalyticsGenerator
 
 _generator = AnalyticsGenerator()
