+++ services/summary_report_generator.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Tuple
 
 import pandas as pd
 
-from .analytics_processor import AnalyticsProcessor
+from .analytics.processor import AnalyticsProcessor
 from .summary_reporter import format_patterns_result
 from .base_database_service import BaseDatabaseService
