+++ tests/test_analysis_extract_utils.py
@@ -1,4 +1,4 @@
-from services.analytics_processing import (
+from analytics.core.utils.results_display import (
     _extract_counts,
     _extract_security_metrics,
 )
