+++ tests/test_enhanced_security_details.py
@@ -1,4 +1,4 @@
-from services.analytics_processing import _extract_enhanced_security_details
+from analytics.core.utils.results_display import _extract_enhanced_security_details
 
 
 def test_extract_enhanced_security_details_basic():
