+++ tests/test_security_display_fix.py
@@ -1,7 +1,7 @@
 import pytest
 import dash_bootstrap_components as dbc
 
-from services.analytics_processing import create_analysis_results_display
+from analytics.core.utils.results_display import create_analysis_results_display
 
 pytestmark = pytest.mark.usefixtures("fake_dbc")
